Fix object size being 0 when retrying request (#14)
Because the code only put the hash in the cache and not the size, the
size that was sent in a retried request was 0.

Some server implementations check the size, so those requests may fail.

To fix it, also put the size in the cache and use it when creating a
request from the cache.